Rotation finished and piece reader component implemented
Introducing the piece reader component which will help read the components of a given piece. No need to reference the piece by name anymore, with this the components become less dependent to each other and only dependent on one glue component. I just need to make it work in the main drag and drop script tho lol