Introduced: Playable Component
Into my first steps to make a playable piece only follow a set movement pattern.

For now a playable piece is only able to move to certain tiles judged by the playable component.

There are still a lot of things to fix, including the tile behavior when its inhabited by a playable piece.