[0.0.3] - Board
[ADDED] :

- TeamsHandler :
Manager for teams with capabilites for
adding new teams and returning existing teams.
- PieceGenerator :
Generator for pieces and setter for playable pieces on playable tiles.
- Board scene :
Container and glue for the TileGrid, PieceGenerator and TeamsHandler components.

- Seeded run :
Now its possible to set a determined seed for every board which every of its components will obey as their flippin god.

- Command : CTRL + SHIFT + R
Reload the current scene to get a new seed.

[CHANGED] :

- TileGrid :
More capabilites for the tilegrid, such as ignoring playable tiles when getting a random tile as well as counting the currently played tiles.

[FIXED] :
- Game hanging when a piece was being randomly assigned to an already played tile.